addr_info: Make init function private.
Move the mptcpd_addr_info_init() to src/commands_upstream.c since that
is the only place it is used, and since it is not intended for use by
the user.
